Update ajax admin

+++ app/Http/Controllers/backend/NhanVienController.php
@@ -9,27 +9,49 @@ use App\Models\LoaiNhanVien;
 use App\http\Requests\NhanVienRequest;
 use Carbon\Carbon;
 
+
 class NhanVienController extends Controller
 {
 
-    public function index()
+    public function index() //danh sách.
     {
         $viewData = [
-            'NhanVien' => NhanVien::orderBy('created_at', 'desc')->get(),
+            'NhanVien' => NhanVien::where('id', '!=', "NV00000000000000")->orderBy('created_at', 'desc')->paginate(10),
             'LoaiNhanVien' => LoaiNhanVien::all(),
         ];
         return view('backend.NhanVien.index', $viewData);
     }
 
-    public function create()
+    public function search(Request $request) // tìm.
+    {
+        $loaiNV = LoaiNhanVien::where([['id', '!=', "LNV00000000000000"], ['tenloainhanvien', $request->search]])->first();
+        if ($loaiNV == null) {
+            $idLoaiNV = "NV00000000000000";
+        } else {
+            $idLoaiNV = $loaiNV->id;
+        }
+        $viewData = [
+            'NhanVien' => NhanVien::where(
+                [['id', '!=', "NV00000000000000"], ['tennhanvien', 'like', '%' . $request->search . '%']]
+            )->orWhere(
+                [['id', '!=', "NV00000000000000"], ['sdt', 'like', '%' . $request->search . '%']]
+            )->orWhere(
+                [['id', '!=', "NV00000000000000"], ['id_loainhanvien', $idLoaiNV]]
+            )->orderBy('created_at', 'desc')->get(),
+            'LoaiNhanVien' => LoaiNhanVien::all(),
+        ];
+        return view('backend.NhanVien.load_NhanVien', $viewData);
+    }
+
+    public function create() // trang thêm. (chưa ajax)
     {
         $viewData = [
-            'LoaiNhanVien' => LoaiNhanVien::where('trangthai', 1)->get(),
+            'LoaiNhanVien' => LoaiNhanVien::where([['id', '!=', "LNV00000000000000"], ['trangthai', 1]])->get(),
         ];
         return view('backend.NhanVien.create_NhanVien', $viewData);
     }
 
-    public function store(NhanVienRequest $request)
+    public function store(NhanVienRequest $request) //thêm. (chưa ajax)
     {
         $iddate = "NV" . Carbon::now('Asia/Ho_Chi_Minh'); //chuỗi thời gian.
         $exp = explode("-", $iddate); //cắt chuỗi.
@@ -46,7 +68,7 @@ class NhanVienController extends Controller
         $data['gioitinh'] = $request->gioitinh;
         $data['luong'] = $request->luong;
         $data['tentaikhoan'] = $request->tentaikhoan;
-        $data['matkhau'] = bcrypt($request->matkhau);
+        $data['password'] = bcrypt($request->password);
         $data['id_loainhanvien'] = $request->id_loainhanvien;
         $data['trangthai'] = $request->trangthai;
         if (empty($request->trangthai)) {
@@ -68,7 +90,7 @@ class NhanVienController extends Controller
         return redirect()->route('nhan-vien.index', $viewData)->with('messsge', "Thêm Thành Công");
     }
 
-    public function show($id)
+    public function show($id) //cho tiết.
     {
         $viewData = [
             'NhanVien' => NhanVien::find($id),
@@ -77,16 +99,16 @@ class NhanVienController extends Controller
         return view('backend.NhanVien.show_NhanVien', $viewData);
     }
 
-    public function edit($id)
+    public function edit($id) //trang cập nhật.  (chưa ajax)
     {
         $viewData = [
             'NhanVien' => NhanVien::find($id),
-            'LoaiNhanVien' => LoaiNhanVien::all(),
+            'LoaiNhanVien' => LoaiNhanVien::where([['id', '!=', "LNV00000000000000"], ['trangthai', 1]])->get(),
         ];
         return view('backend.NhanVien.edit_NhanVien', $viewData);
     }
 
-    public function update(NhanVienRequest $request, $id)
+    public function update(NhanVienRequest $request, $id) //cập nhật.  (chưa ajax)
     {
         $data['tennhanvien'] = ucwords($request->tennhanvien);
         $data['sdt'] = $request->sdt;
@@ -102,11 +124,11 @@ class NhanVienController extends Controller
             $data['trangthai'] = 0;
         }
         $OldNhanVien = NhanVien::find($id); //lấy nhân viên củ.
-        if ($OldNhanVien->matkhau != $request->matkhau) { //kiểm tra có thay đổ mật khẩu không.
-            $data['matkhau'] = bcrypt($request->matkhau);
+        if ($OldNhanVien->password != $request->password) { //kiểm tra có thay đổ mật khẩu không.
+            $data['password'] = bcrypt($request->password);
         }
         if ($request->hasFile('hinhanh')) { //kiểm tra xem có file không.
-            if ($OldNhanVien->hinhanh != "NV.png") { //kiểm tra có phải đang dùng hình ảnh mặc định không.
+            if ($OldNhanVien->hinhanh != "NOIMAGE.png") { //kiểm tra có phải đang dùng hình ảnh mặc định không.
                 unlink('uploads/NhanVien/' . $OldNhanVien->hinhanh); //xoá ảnh củ.
             }
             $file = $request->hinhanh; //lấy tên hình được gửi lên.
@@ -123,15 +145,15 @@ class NhanVienController extends Controller
         return redirect()->route('nhan-vien.index', $viewData)->with('messsge', "Cập Nhật Thành Công");
     }
 
-    public function destroy($id)
+    public function destroy($id) // xóa.
     {
         $OldNhanVien = NhanVien::find($id); //lấy nhân viên củ
-        unlink('uploads/NhanVien/' . $OldNhanVien->hinhanh); //xoá ảnh củ.
-        NhanVien::where('id', $id)->delete();
-        $viewData = [
-            'NhanVien' => NhanVien::orderBy('created_at', 'desc')->get(),
-            'LoaiNhanVien' => LoaiNhanVien::all(),
-        ];
-        return redirect()->route('nhan-vien.index', $viewData)->with('messsge', "Xoá Thành Công");
+        if ($OldNhanVien->hinhanh != "NOIMAGE.png") { //kiểm tra có phải đang dùng hình ảnh mặc định không.
+            unlink('uploads/NhanVien/' . $OldNhanVien->hinhanh); //xoá ảnh củ.
+        }
+        NhanVien::where(
+            [['id', '!=', "NV00000000000000"], ['id', $id]]
+        )->delete();
+        return response()->json(['success' => 'Thành Công Rồi']);
     }
 }
